+++ src/ZLinq/Linq/Min.cs
@@ -206,7 +206,7 @@ partial class ValueEnumerableExtensions
             }
             while (value is null);
 
-            while (index < span.Length)
+            while ((uint)index < (uint)span.Length)
             {
                 // compare both(left, right) non-null
                 if (span[index] is not null && comparer.Compare(span[index], value) < 0)
@@ -231,7 +231,7 @@ partial class ValueEnumerableExtensions
             // optimize for default comparer
             if (comparer == Comparer<TSource>.Default)
             {
-                while (index < span.Length)
+                while ((uint)index < (uint)span.Length)
                 {
                     if (Comparer<TSource>.Default.Compare(span[index], value) < 0)
                     {
@@ -244,7 +244,7 @@ partial class ValueEnumerableExtensions
             }
             else
             {
-                while (index < span.Length)
+                while ((uint)index < (uint)span.Length)
                 {
                     if (comparer.Compare(span[index], value) < 0)
                     {
